Changed main GUI File to application GUI and seperated serial monitor code from application code
